requirements.txt
kivy==2.0.0
plyer==2.0.0
python-dotenv==0.19.1